requirements.txt
asgiref==3.7.2
crispy-bootstrap4==2023.1
Django==4.2.6
django-crispy-forms==2.1
mysqlclient==2.2.1
Pillow==10.0.1
sqlparse==0.4.4
typing_extensions==4.8.0
tzdata==2023.3
